Move Sprite enum to model
It will be used by the new DesktopModel class.
